requirements/production.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.2
ecdsa==0.17.0
fastapi==0.75.2
h11==0.13.0
idna==3.3
peewee==3.14.10
peewee-migrate==1.4.7
pyasn1==0.4.8
pydantic==1.9.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.2.0
uvicorn==0.17.6
